- Did the shader - Outdoor and Indoor works with placing and digging - There is a massive bug when marking neighbor blocks dirty that I added some debug with squirrel to try and track down but couldn't fix it

+++ Engine/Code/Engine/Renderer/Renderer.cpp
@@ -1386,7 +1386,7 @@ void Renderer::EnableWireframe(bool check)
 //-----------------------------------------------------------------------------------------------
 void Renderer::SetPointSize(int size)
 {
-	glPointSize(size);
+	glPointSize((GLfloat) size);
 }
 
 Texture* Renderer::CreateRenderTarget(int width, int height, eTextureFormat fmt /*= TEXTURE_FORMAT_RGBA8 */)
